+++ src/plaid/utils/cgns_worker.py
@@ -16,7 +16,9 @@ import CGNS.MAP as CGM
 
 logger = logging.getLogger(__name__)
 
-if __name__ == "__main__":
+if (
+    __name__ == "__main__"
+):  # pragma: no cover (run in subprocess, not picked up by coverage routine)
     tmpfile, outfname = sys.argv[1], sys.argv[2]
     with open(tmpfile, "rb") as f:
         tree = pickle.load(f)